complete the rename
