+++ Code/GraphMol/MolInteractionFields/Wrap/testData/aceticacid.mol
