+++ Code/GraphMol/MolInteractionFields/test_data/aceticacid.mol
