+++ Code/GraphMol/MolInteractionFields/test_data/alicin.mol
