+++ Code/GraphMol/MolInteractionFields/test_data/ammonia.mol
