+++ Code/GraphMol/MolInteractionFields/test_data/diethylether.mol
