+++ Code/GraphMol/MolInteractionFields/test_data/glucose.mol
