+++ Code/GraphMol/MolInteractionFields/test_data/glycine.mol
