+++ Code/GraphMol/MolInteractionFields/test_data/glythrgly.mol
