+++ Code/GraphMol/MolInteractionFields/test_data/h2o.mol
