+++ Code/GraphMol/MolInteractionFields/test_data/methylammonium.mol
