+++ Code/GraphMol/MolInteractionFields/test_data/serine.mol
