+++ Code/GraphMol/MolInteractionFields/test_data/sulfanilamide.mol
