+++ Code/GraphMol/MolInteractionFields/test_data/threonine.mol
